+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/bug/IncorrectOverriddenForDefinitionInspection.kt
@@ -14,16 +14,16 @@ import icu.windea.pls.lang.search.selector.*
 import icu.windea.pls.lang.util.*
 import icu.windea.pls.model.*
 import icu.windea.pls.script.psi.*
+import org.jetbrains.annotations.Nls
 
 /**
  * （对于脚本文件）检查是否存在不正确的对定义的重载。
  */
 class IncorrectOverriddenForDefinitionInspection : LocalInspectionTool() {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        if (!shouldCheckFile(holder.file)) return PsiElementVisitor.EMPTY_VISITOR
-
         val file = holder.file
         val project = holder.project
+        if (!shouldCheckFile(file)) return PsiElementVisitor.EMPTY_VISITOR
         val fileInfo = file.fileInfo ?: return PsiElementVisitor.EMPTY_VISITOR
         val virtualFile = file.virtualFile
         val inProject = virtualFile != null && ProjectFileIndex.getInstance(project).isInContent(virtualFile)
@@ -44,19 +44,20 @@ class IncorrectOverriddenForDefinitionInspection : LocalInspectionTool() {
                 val selector = selector(project, file).definition()
                 val name = definitionInfo.name
                 val type = definitionInfo.type
-                if (name.isParameterized()) return ////parameterized -> ignored
+                if (name.isEmpty()) return //anonymous -> skipped
+                if (name.isParameterized()) return //parameterized -> ignored
                 val results = ParadoxDefinitionSearch.search(name, type, selector).findAll()
                 if (results.size < 2) return //no override -> skip
                 val firstResult = results.first()
                 val firstRootInfo = firstResult.fileInfo?.rootInfo ?: return
                 val rootInfo = fileInfo.rootInfo
-                if (firstRootInfo.rootFile != rootInfo.rootFile) {
-                    //different root file -> incorrect override
-                    val locationElement = element.propertyKey
-                    val message = PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.desc", name, priority)
-                    val fix = NavigateToOverriddenDefinitionsFix(name, element, results)
-                    holder.registerProblem(locationElement, message, fix)
-                }
+                if (firstRootInfo.rootFile == rootInfo.rootFile) return
+
+                //different root file -> incorrect override
+                val locationElement = element.propertyKey
+                val message = PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.desc", name, priority)
+                val fix = NavigateToOverriddenDefinitionsFix(name, element, results)
+                holder.registerProblem(locationElement, message, fix)
             }
         }
     }
@@ -70,10 +71,13 @@ class IncorrectOverriddenForDefinitionInspection : LocalInspectionTool() {
     private class NavigateToOverriddenDefinitionsFix(key: String, element: PsiElement, definitions: Collection<PsiElement>) : NavigateToFix(key, element, definitions) {
         override fun getText() = PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.fix.1")
 
-        override fun getPopupTitle(editor: Editor) =
-            PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.fix.1.popup.title", key)
+        override fun getPopupTitle(editor: Editor): @Nls String {
+            return PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.fix.1.popup.title", key)
+        }
 
-        override fun getPopupText(editor: Editor, value: PsiElement) =
-            PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.fix.1.popup.text", key, editor.document.getLineNumber(value.textOffset))
+        override fun getPopupText(editor: Editor, value: PsiElement): @Nls String {
+            val lineNumber = editor.document.getLineNumber(value.textOffset)
+            return PlsBundle.message("inspection.script.incorrectOverriddenForDefinition.fix.1.popup.text", key, lineNumber)
+        }
     }
 }