+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/bug/IncorrectOverriddenForScriptedVariableInspection.kt
@@ -14,16 +14,16 @@ import icu.windea.pls.lang.search.*
 import icu.windea.pls.lang.search.selector.*
 import icu.windea.pls.lang.util.*
 import icu.windea.pls.script.psi.*
+import org.jetbrains.annotations.Nls
 
 /**
  * （对于脚本文件）检查是否存在不正确的对（全局）封装的重载。
  */
 class IncorrectOverriddenForScriptedVariableInspection : LocalInspectionTool() {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        if (!shouldCheckFile(holder.file)) return PsiElementVisitor.EMPTY_VISITOR
-
         val file = holder.file
         val project = holder.project
+        if (!shouldCheckFile(file)) return PsiElementVisitor.EMPTY_VISITOR
         val fileInfo = file.fileInfo ?: return PsiElementVisitor.EMPTY_VISITOR
         val isGlobal = "common/scripted_variables".matchesPath(fileInfo.path.path)
         if (!isGlobal) return PsiElementVisitor.EMPTY_VISITOR //only for global scripted variables
@@ -44,20 +44,21 @@ class IncorrectOverriddenForScriptedVariableInspection : LocalInspectionTool() {
                 val priority = ParadoxPriorityProvider.getPriority(element)
                 if (priority == ParadoxPriority.ORDERED) return //only for FIOS and LIOS
                 val selector = selector(project, file).scriptedVariable()
-                val name = element.name ?: return
-                if (name.isParameterized()) return ////parameterized -> ignored
+                val name = element.name
+                if (name.isNullOrEmpty()) return //anonymous -> skipped
+                if (name.isParameterized()) return //parameterized -> ignored
                 val results = ParadoxGlobalScriptedVariableSearch.search(name, selector).findAll()
                 if (results.size < 2) return //no override -> skip
                 val firstResult = results.first()
                 val firstRootInfo = firstResult.fileInfo?.rootInfo ?: return
                 val rootInfo = fileInfo.rootInfo
-                if (firstRootInfo.rootFile != rootInfo.rootFile) {
-                    //different root file -> incorrect override
-                    val locationElement = element.scriptedVariableName
-                    val message = PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.desc", name, priority)
-                    val fix = NavigateToOverriddenScriptedVariablesFix(name, element, results)
-                    holder.registerProblem(locationElement, message, fix)
-                }
+                if (firstRootInfo.rootFile == rootInfo.rootFile) return
+
+                //different root file -> incorrect override
+                val locationElement = element.scriptedVariableName
+                val message = PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.desc", name, priority)
+                val fix = NavigateToOverriddenScriptedVariablesFix(name, element, results)
+                holder.registerProblem(locationElement, message, fix)
             }
         }
     }
@@ -71,11 +72,14 @@ class IncorrectOverriddenForScriptedVariableInspection : LocalInspectionTool() {
     private class NavigateToOverriddenScriptedVariablesFix(key: String, element: PsiElement, scriptedVariables: Collection<PsiElement>) : NavigateToFix(key, element, scriptedVariables) {
         override fun getText() = PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.fix.1")
 
-        override fun getPopupTitle(editor: Editor) =
-            PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.fix.1.popup.title", key)
+        override fun getPopupTitle(editor: Editor): @Nls String {
+            return PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.fix.1.popup.title", key)
+        }
 
-        override fun getPopupText(editor: Editor, value: PsiElement) =
-            PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.fix.1.popup.text", key, editor.document.getLineNumber(value.textOffset))
+        override fun getPopupText(editor: Editor, value: PsiElement): @Nls String {
+            val lineNumber = editor.document.getLineNumber(value.textOffset)
+            return PlsBundle.message("inspection.script.incorrectOverriddenForScriptedVariable.fix.1.popup.text", key, lineNumber)
+        }
     }
 }