+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/common/OverriddenForScriptedVariableInspection.kt
@@ -12,16 +12,16 @@ import icu.windea.pls.lang.quickfix.*
 import icu.windea.pls.lang.search.*
 import icu.windea.pls.lang.search.selector.*
 import icu.windea.pls.script.psi.*
+import org.jetbrains.annotations.Nls
 
 /**
  * （对于脚本文件）检查是否存在对（全局）封装变量的重载。
  */
 class OverriddenForScriptedVariableInspection : LocalInspectionTool() {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        if (!shouldCheckFile(holder.file)) return PsiElementVisitor.EMPTY_VISITOR
-
         val file = holder.file
         val project = holder.project
+        if (!shouldCheckFile(file)) return PsiElementVisitor.EMPTY_VISITOR
         val fileInfo = file.fileInfo ?: return PsiElementVisitor.EMPTY_VISITOR
         val isGlobal = "common/scripted_variables".matchesPath(fileInfo.path.path)
         if (!isGlobal) return PsiElementVisitor.EMPTY_VISITOR //only for global scripted variables
@@ -40,7 +40,8 @@ class OverriddenForScriptedVariableInspection : LocalInspectionTool() {
 
             private fun visitScriptedVariable(element: ParadoxScriptScriptedVariable) {
                 val selector = selector(project, file).scriptedVariable()
-                val name = element.name ?: return
+                val name = element.name
+                if (name.isNullOrEmpty()) return //anonymous -> skipped
                 if (name.isParameterized()) return //parameterized -> ignored
                 val results = ParadoxGlobalScriptedVariableSearch.search(name, selector).findAll()
                 if (results.size < 2) return //no override -> skip
@@ -61,11 +62,14 @@ class OverriddenForScriptedVariableInspection : LocalInspectionTool() {
     private class NavigateToOverriddenScriptedVariablesFix(key: String, element: PsiElement, elements: Collection<PsiElement>) : NavigateToFix(key, element, elements) {
         override fun getText() = PlsBundle.message("inspection.script.overriddenForScriptedVariable.fix.1")
 
-        override fun getPopupTitle(editor: Editor) =
-            PlsBundle.message("inspection.script.overriddenForScriptedVariable.fix.1.popup.title", key)
+        override fun getPopupTitle(editor: Editor): @Nls String {
+            return PlsBundle.message("inspection.script.overriddenForScriptedVariable.fix.1.popup.title", key)
+        }
 
-        override fun getPopupText(editor: Editor, value: PsiElement) =
-            PlsBundle.message("inspection.script.overriddenForScriptedVariable.fix.1.popup.text", key, editor.document.getLineNumber(value.textOffset))
+        override fun getPopupText(editor: Editor, value: PsiElement): @Nls String {
+            val lineNumber = editor.document.getLineNumber(value.textOffset)
+            return PlsBundle.message("inspection.script.overriddenForScriptedVariable.fix.1.popup.text", key, lineNumber)
+        }
     }
 }